rename deativate-all-other-plugins + optimize
- skip navigation if already on plugins.php
- skip bulk action if site kit is the only plugin active
