+++ tests/e2e/utils/index.js
@@ -1,3 +1,3 @@
-export { deactivateAllOtherPlugins } from './deactivateAllOtherPlugins';
+export { deactivateAllOtherPlugins } from './deactivate-all-other-plugins';
 export { pasteText } from './paste-text';
 export { resetSiteKit } from './reset';
